Ajout des methodes des autres classes

+++ src/main/java/com/inti/entities/Hotel.java
@@ -9,7 +9,7 @@ import javax.persistence.Entity;
 @DiscriminatorValue("hotel")
 public class Hotel extends Lieu implements Serializable {
 	private Float proprete;
-	private Long nbEtoiles;
+	private int nbEtoiles;
 	private Float prixNuit;
 	
 	public Hotel() {
@@ -21,7 +21,7 @@ public class Hotel extends Lieu implements Serializable {
 	public Hotel(String nom, String localisation, String nbVisiteurs) {
 		super(nom, localisation, nbVisiteurs);
 	}
-	public Hotel(Float proprete, Long nbEtoiles, Float prixNuit) {
+	public Hotel(Float proprete, int nbEtoiles, Float prixNuit) {
 		super();
 		this.proprete = proprete;
 		this.nbEtoiles = nbEtoiles;
@@ -33,10 +33,10 @@ public class Hotel extends Lieu implements Serializable {
 	public void setProprete(Float proprete) {
 		this.proprete = proprete;
 	}
-	public Long getNbEtoiles() {
+	public int getNbEtoiles() {
 		return nbEtoiles;
 	}
-	public void setNbEtoiles(Long nbEtoiles) {
+	public void setNbEtoiles(int nbEtoiles) {
 		this.nbEtoiles = nbEtoiles;
 	}
 	public Float getPrixNuit() {
